modify:调整im监听代码相关逻辑位置.
