+++ src/App.vue
@@ -1,211 +1,14 @@
 <script setup>
-import { useStore } from 'vuex'
-import router from '@/router'
+import { ref } from 'vue'
+import { mountAllEMListener } from '@/IM/listener'
 import { EaseChatSDK, EaseChatClient } from '@/IM/initwebsdk'
-import { handleSDKErrorNotifi, setMessageKey } from '@/utils/handleSomeData'
-import { informType, messageType } from '@/constant'
-import { usePlayRing } from '@/hooks'
 import ring from '@/assets/ring.mp3'
 /* callkit */
 import EaseCallKit from '@/components/EaseCallKit'
 import InviteCallMembers from '@/components/InviteCallMembers'
-import { ref } from 'vue'
-const store = useStore()
-const { isOpenPlayRing, clickRing } = usePlayRing()
+
 EaseChatSDK.logger.disableAll()
-/* connect 相关监听 */
-EaseChatClient.addEventHandler('connection', {
-    onConnected: () => {
-        console.log('>>>>>环信连接成功')
-        store.commit('CHANGE_LOGIN_STATUS', true)
-        if (isOpenPlayRing.value) clickRing()
-        fetchLoginUsersInitData()
-        router.replace('/chat')
-    },
-    onDisconnected: () => {
-        router.push('/login')
-        store.commit('CHANGE_LOGIN_STATUS', false)
-    },
-    onOnline: () => {
-        store.commit('CHANGE_NETWORK_STATUS', true)
-    }, // 本机网络连接成功。
-    onOffline: () => {
-        store.commit('CHANGE_NETWORK_STATUS', false)
-    }, // 本机网络掉线。
-    onError: (error) => {
-        console.log('on error', error)
-        handleSDKErrorNotifi(error.type, error.message)
-    }
-})
-//fetch 登陆用户的初始数据
-const fetchLoginUsersInitData = () => {
-    getMyUserInfos()
-    fetchFriendList()
-    fetchTheLoginUserBlickList()
-    fetchGroupList()
-    //初始化vuex中的会话列表相关数据
-    store.commit('INIT_CONVERSATION_STATE')
-}
-//获取登陆用户属性
-const getMyUserInfos = () => {
-    const userId = EaseChatClient.user
-    store.dispatch('getMyUserInfo', userId)
-}
-//获取好友列表
-const fetchFriendList = () => {
-    // const { value = {} } = useLocalStorage('friendList')
-    // if (Object.values(JSON.parse(value)).length > 0) return
-    store.dispatch('fetchAllFriendListFromServer')
-}
-//获取黑名单列表
-const fetchTheLoginUserBlickList = () => store.dispatch('fetchBlackList')
-//获取加入的群组列表
-const fetchGroupList = () => {
-    //如果本地存储里不存在群组列表则调用获取群组列表
-    // const { value = {} } = useLocalStorage('groupList')
-    // if (Object.values(JSON.parse(value)).length > 0) return
-    const pageParams = {
-        pageNum: 1,
-        pageSize: 20
-    }
-    store.dispatch('fetchGroupList', pageParams)
-}
-
-/* presence 相关监听 */
-EaseChatClient.addEventHandler('presenceStatusChange', {
-    onPresenceStatusChange: (status) => {
-        console.log('>>>>>presenceStatusChange', status)
-        getUserPresence(...status)
-    }
-})
-//处理登陆用户状态的变更
-const getUserPresence = (status) => {
-    store.dispatch('handlePresenceChanges', status)
-}
-const { CHANGE_MESSAGE_BODAY_TYPE } = messageType
-/* message 相关监听 */
-EaseChatClient.addEventHandler('messageListen', {
-    onTextMessage: function (message) {
-        console.log('>>>>>>>App mesage', message)
-        console.log('setMessageKey', setMessageKey(message))
-        pushNewMessage(message)
-    }, // 收到文本消息。
-    onEmojiMessage: function (message) {
-        pushNewMessage(message)
-    }, // 收到表情消息。
-    onImageMessage: function (message) {
-        pushNewMessage(message)
-    }, // 收到图片消息。
-    onCmdMessage: function (message) {
-        console.log('>>>>>收到命令消息', message)
-    }, // 收到命令消息。
-    onAudioMessage: function (message) {
-        pushNewMessage(message)
-    }, // 收到音频消息。
-    onLocationMessage: function (message) {
-        pushNewMessage(message)
-    }, // 收到位置消息。
-    onFileMessage: function (message) {
-        pushNewMessage(message)
-    }, // 收到文件消息。
-    onCustomMessage: function (message) {
-        pushNewMessage(message)
-    }, // 收到自定义消息。
-    onVideoMessage: function (message) {
-        pushNewMessage(message)
-    }, // 收到视频消息。
-    onRecallMessage: function (message) {
-        otherRecallMessage(message)
-    }, // 收到消息撤回回执。
-    onModifiedMessage: function (message) {
-        console.log('>>>>>收到修改消息', message)
-        otherModifyMessage(message)
-    }
-})
-//接收的消息往store中push
-const pushNewMessage = (message) => {
-    store.dispatch('createNewMessage', message)
-}
-//收到他人的撤回指令
-const otherRecallMessage = (message) => {
-    console.log('>>>>>收到他人撤回', message)
-    const { from, to, mid } = message
-    //单对单的撤回to必然为登陆的用户id，群组发起撤回to必然为群组id 所以key可以这样来区分群组或者单人。
-    const key = to === EaseChatClient.user ? from : to
-    console.log('>>>>>收到他人撤回', key)
-    store.commit('CHANGE_MESSAGE_BODAY', {
-        type: CHANGE_MESSAGE_BODAY_TYPE.RECALL,
-        key,
-        mid
-    })
-    store.dispatch('gatherConversation', key)
-}
-//收到消息修改指令
-const otherModifyMessage = (message) => {
-    const { from, to, id: mid } = message
-    //单对单的撤回to必然为登陆的用户id，群组发起撤回to必然为群组id 所以key可以这样来区分群组或者单人。
-    if (!to) return
-    const key = to === EaseChatClient.user ? from : to
-    store.commit('CHANGE_MESSAGE_BODAY', {
-        type: CHANGE_MESSAGE_BODAY_TYPE.MODIFY,
-        key,
-        mid,
-        message
-    })
-    store.dispatch('gatherConversation', key)
-}
-/* 好友关系相关监听 */
-const { INFORM_FROM } = informType
-EaseChatClient.addEventHandler('friendListen', {
-    // 收到好友邀请触发此方法。
-    onContactInvited: (data) => {
-        //写入INFORM
-        console.log('>>>>>>收到好友申请', data)
-        submitInformData(INFORM_FROM.FRIEND, data)
-    },
-    // 联系人被删除时触发此方法。
-    onContactDeleted: (data) => {
-        //写入INFORM
-        console.log('>>>>收到好友关系解散', data)
-        submitInformData(INFORM_FROM.FRIEND, data)
-        store.dispatch('onDeleteFriend', data)
-    },
-    // 新增联系人会触发此方法。
-    onContactAdded: (data) => {
-        console.log('>>>>好友新增监听', data)
-        submitInformData(INFORM_FROM.FRIEND, data)
-        store.dispatch('onAddedNewFriend', data)
-    },
-    // 好友请求被拒绝时触发此方法。
-    onContactRefuse: (data) => {
-        //写入INFORM
-        console.log('>>>>>>好友申请被拒绝', data)
-        data.type = 'other_person_refuse'
-        submitInformData(INFORM_FROM.FRIEND, data)
-    },
-    // 好友请求被同意时触发此方法。
-    onContactAgreed: (data) => {
-        //写入INFORM
-        console.log('>>>>>对方同意了好友申请', data)
-        //改掉data中的type
-        data.type = 'other_person_agree'
-        submitInformData(INFORM_FROM.FRIEND, data)
-        store.dispatch('onAddedNewFriend', data)
-    }
-})
-/* 群组相关监听 */
-EaseChatClient.addEventHandler('groupEvent', {
-    onGroupEvent: (groupevent) => {
-        console.log('>>>>>>>收到群组事件', groupevent)
-        submitInformData(INFORM_FROM.GROUP, groupevent)
-    }
-})
-
-const submitInformData = (fromType, informContent) => {
-    console.log('>>>submitInformData>>>', fromType, informContent)
-    store.dispatch('createNewInform', { fromType, informContent })
-}
-
+mountAllEMListener()
 /* 重新登陆 */
 //读取本地EASEIM_loginUser
 const EASEIM_loginUser = window.localStorage.getItem('EASEIM_loginUser')
