+++ src/IM/config/index.js
@@ -1,0 +1,4 @@
+//环信appKey默认配置项
+export const DEFAULT_EASEMOB_APPKEY = 'easemob#easeim'
+export const DEFAULT_EASEMOB_SOCKET_URL = '//im-api-v2.easemob.com/ws'
+export const DEFAULT_EASEMOB_REST_URL = '//a1.easemob.com'
